requirements.txt
discord-webhook~=0.9.0
requests~=2.24.0
python-dateutil~=2.8.1
